+++ src/main/java/legend/game/combat/Bttl_800c.java
@@ -41,6 +41,7 @@ import legend.game.combat.effects.BttlScriptData6cSub13c;
 import legend.game.combat.effects.EffectManagerData6c;
 import legend.game.combat.effects.FullScreenOverlayEffect0e;
 import legend.game.combat.effects.RadialGradientEffect14;
+import legend.game.combat.effects.SpriteMetrics08;
 import legend.game.combat.effects.WeaponTrailEffect3c;
 import legend.game.combat.effects.WeaponTrailEffectSegment2c;
 import legend.game.combat.environment.BattleCamera;
@@ -56,7 +57,6 @@ import legend.game.combat.types.CombatantStruct1a8;
 import legend.game.combat.types.CombatantStruct1a8_c;
 import legend.game.combat.types.DragoonSpells09;
 import legend.game.combat.types.MersenneTwisterSeed;
-import legend.game.combat.types.SpriteMetrics08;
 import legend.game.combat.ui.BattleDisplayStats144;
 import legend.game.combat.ui.BattleMenuStruct58;
 import legend.game.combat.ui.BattleStruct3c;
