+++ src/main/java/legend/game/combat/SEffe.java
@@ -50,23 +50,9 @@ import legend.game.combat.effects.AdditionOverlaysEffect44;
 import legend.game.combat.effects.AdditionOverlaysHit20;
 import legend.game.combat.effects.AttackHitFlashEffect0c;
 import legend.game.combat.effects.BattleStruct24;
-import legend.game.combat.effects.StarChildrenMeteorEffect10;
 import legend.game.combat.effects.BttlScriptData6cSub13c;
-import legend.game.combat.effects.WsDragoonTransformationFeathersEffect14;
-import legend.game.combat.effects.WsDragoonTransformationFeatherInstance70;
-import legend.game.combat.effects.MoonlightStarsEffect18;
-import legend.game.combat.effects.MoonlightStar3c;
-import legend.game.combat.effects.ThunderArrowEffect1c;
-import legend.game.combat.effects.ThunderArrowEffectBolt1e;
 import legend.game.combat.effects.BttlScriptData6cSub1c_3;
-import legend.game.combat.effects.StarChildrenImpactEffect20;
 import legend.game.combat.effects.BttlScriptData6cSub20_2Suba8;
-import legend.game.combat.effects.GradientRaysEffect24;
-import legend.game.combat.effects.UnusedBuggedEffect24;
-import legend.game.combat.effects.SpriteWithTrailEffect30;
-import legend.game.combat.effects.SpriteWithTrailEffect30Sub10;
-import legend.game.combat.effects.LensFlareEffect50;
-import legend.game.combat.effects.LensFlareEffectInstance3c;
 import legend.game.combat.effects.BttlScriptData6cSub5c;
 import legend.game.combat.effects.BttlScriptData6cSubBase1;
 import legend.game.combat.effects.DeffTmdRenderer14;
@@ -78,9 +64,14 @@ import legend.game.combat.effects.ElectricityEffect38;
 import legend.game.combat.effects.FrozenJetEffect28;
 import legend.game.combat.effects.GoldDragoonTransformEffect20;
 import legend.game.combat.effects.GoldDragoonTransformEffectInstance84;
+import legend.game.combat.effects.GradientRaysEffect24;
+import legend.game.combat.effects.LensFlareEffect50;
+import legend.game.combat.effects.LensFlareEffectInstance3c;
 import legend.game.combat.effects.LightningBoltEffect14;
 import legend.game.combat.effects.LightningBoltEffectSegment30;
 import legend.game.combat.effects.LightningBoltEffectSegmentOrigin08;
+import legend.game.combat.effects.MoonlightStar3c;
+import legend.game.combat.effects.MoonlightStarsEffect18;
 import legend.game.combat.effects.ParticleEffectData98;
 import legend.game.combat.effects.ParticleEffectInstance94;
 import legend.game.combat.effects.ParticleEffectInstance94Sub10;
@@ -89,12 +80,21 @@ import legend.game.combat.effects.RaindropEffect0c;
 import legend.game.combat.effects.ScreenCaptureEffect1c;
 import legend.game.combat.effects.ScreenCaptureEffectMetrics8;
 import legend.game.combat.effects.ScreenDistortionEffectData08;
+import legend.game.combat.effects.SpriteMetrics08;
+import legend.game.combat.effects.SpriteWithTrailEffect30;
+import legend.game.combat.effects.SpriteWithTrailEffect30Sub10;
+import legend.game.combat.effects.StarChildrenImpactEffect20;
+import legend.game.combat.effects.StarChildrenMeteorEffect10;
+import legend.game.combat.effects.ThunderArrowEffect1c;
+import legend.game.combat.effects.ThunderArrowEffectBolt1e;
 import legend.game.combat.effects.TransformScalerEffect34;
+import legend.game.combat.effects.UnusedBuggedEffect24;
+import legend.game.combat.effects.WsDragoonTransformationFeatherInstance70;
+import legend.game.combat.effects.WsDragoonTransformationFeathersEffect14;
 import legend.game.combat.types.BattleScriptDataBase;
 import legend.game.combat.types.DragoonAdditionScriptData1c;
 import legend.game.combat.types.EffeScriptData30;
 import legend.game.combat.types.EffeScriptData30Sub06;
-import legend.game.combat.types.SpriteMetrics08;
 import legend.game.combat.types.VertexDifferenceAnimation18;
 import legend.game.combat.ui.AdditionOverlayMode;
 import legend.game.modding.coremod.CoreMod;
