+++ src/main/java/legend/game/combat/deff/DeffManager7cc.java
@@ -2,10 +2,10 @@ package legend.game.combat.deff;
 
 import legend.core.gte.TmdObjTable1c;
 import legend.game.combat.effects.EffectManagerData6c;
+import legend.game.combat.effects.SpriteMetrics08;
 import legend.game.combat.environment.BattleLightStruct64;
 import legend.game.combat.environment.BttlLightStruct84;
 import legend.game.combat.environment.StageAmbiance4c;
-import legend.game.combat.types.SpriteMetrics08;
 import legend.game.scripting.ScriptFile;
 import legend.game.scripting.ScriptState;
