+++ src/main/java/legend/game/combat/effects/AttackHitFlashEffect0c.java
@@ -1,7 +1,5 @@
 package legend.game.combat.effects;
 
-import legend.game.combat.types.SpriteMetrics08;
-
 public class AttackHitFlashEffect0c extends SpriteWithTrailEffect30.Sub implements BttlScriptData6cSubBase1 {
   public final SpriteMetrics08 metrics_04 = new SpriteMetrics08();
 }
